Refactored Stylesheets
style: Added custom mixin to load themes cleanly.

Added a Theme folder for all theme partials to be broken down and
imported as needed, as well as, wrote a custom mixin 'load-theme( )'
which cleans up the code a bit and prevents repetition.  I also changed
the font-weight of the paragraph text to 300 for better readability as
the current weight was very heavy.