refactor: specialist changes stats instead of adding bonus
After the specialist bonus was introduced, there was some confusion
since it affected the bonus instead of the actual stats. Since the soul
is an equippable item, it makes more sense to reflect the change in the
stats directly. The specialist checkbox will now update the base stats
in the simulator and profile instead of just adding a bonus in the sim.
Also refactored a shadowed variable in the stats popup.