feat: quick commission button on search results
Adds a commission button to the Items search that allows a user to
quickly create a commission for that single item. Updates the commission
dialog to allow suppression of the list warning since it does not apply
in this case. Updates the flow for deleting a commission so that the
list is destroyed if it was a quick commission.

Resolves #490