refactor: new custom rotation button to new rotation
Currently a lot of new users press the "new custom rotation" button when
they really want to create a rotation for an existing recipe. To solve
this, the New button on the Rotations page now opens the recipe search
dialog. If the user wishes to create a custom rotation, they can select
that option from the bottom of the dialog.