refactor: pricing mode overhaul
The cost for pre crafts is now determined automatically by the total
crafting cost of their recipes. A checkbox has been added that, when
selected, allows the user to override this value with a custom price.
This setting is persisted across all lists using local storage.

The total amount spent is now determined strictly by the crafting cost
for the final products' recipes, instead of the sum of all categories.
As such, the summed values have been removed from the individual panels
since they are no longer used in the calculation. The "materials have
cost" warning has also been removed since this is no longer relevant
with the new calculation.