feat: commission board history
Adds another tab to the commission board where users can view the
previous commissions archived for their server. Updates the commission
panel to hide the Archived tag, since it is redundant. Updates the
archived commission fetch to query the status from Firestore and sort
the results in descending date order.

Resolves #519