feat: convert custom rotation to normal rotation
Adds a button to the Simulator page that allows users to convert a saved
custom rotation into a normal rotation. This is done by allowing the
user to search for a recipe to use for the rotation, and then saving
that item/recipe with the rotation so it is opened like a normal one.

Resolves #609