feat: levequests
Adds the ability for users to search for Levequests. All relevant
Levequest information is displayed, including the requested item,
location, rewards, etc. Users can then add a specified quantity of the
Levequest to a list. This operation factors in the quantity desired,
the amount required by the Levequest, and 3x turn-in if selected.

Also adds a new pipe, XivapiIcon, for displaying icons from XIVAPI.