feat(simulator): Add consumables to generated macros (#717)
Since various macros created via the crafting rotations UI can use very
different consumables (food, tea), it can get cumbersome to remember
which macro uses what combination. This changeset adds a new line to the
generated macro for cross class action setup which informs the user what
food was used in the macro. The effect would be that, when a user is
about to use a macro, they'll have the food expected to be used splatted
out in the console