Optimizer panels no longer hide when empty
For optimizer panels with interactive input fields on them, if they are
hidden due to the filters being too strict, that makes it so that the
user will not be able to alter the values.

To avoid this, the panels are no longer hidden if they are empty. The
user can refer to the count presented on the panel text itself to
determine if they should ignore it or not (e.g. zero items).