Added "Consolidate Stacks" Inventory Optimizer
A new inventory optimizer that will suggest high-quality (HQ) stacks of
items that can have their quality lowered so that it can be joined with
an existing normal-quality (NQ) stack of that same type of item. This is
a very reliable way to free up inventory slots.

Because users may not want all stacks of items suggested, filtering is
provided by expansion. For example, if a user wants all items in
Heavensward and below suggested (Heavensward + A Realm Reborn), they
will select "Heavensward" from the filter select box. Items with HQ and
NQ stacks from Stormblood and Shadowbringers will *not* be selected in
this case because they are filtered out. You can also clear the filter
box (so it says "Pick Expansion") which completely disables filtering,
showing all available items in your inventory to which the optimization
applies.

In general, if item database information is incomplete, an item does not
get filtered even if it may have been otherwise. Examples:

* The specified item does not have an assigned patch.
* A patch to which an item is assigned is not defined.